Probably not the recommended thing to do and, if things do change, maintanence is a pain. But it wasn't terribly difficult to do, with a permutations calculator and Emacs macros.
And, honestly, it's a Hell of a lot nicer user experience to be able
to just create the object with the exact items you want rather than
having to remember to put `null`; I suppose the later forces the user
to keep in mind what they're passing and what their skipping but not
enough for me to care (plus, if you want to pass a particular
argument, chances are you won't forget to).

I need the change the `Object` array once I figure out what's actually
in the array but I can do that later.

I sort of wanted just a general `Post` class that held all the
possible instance variables and, then, each extending class could just
specify their own constructors and getters to set what's accessible.

But couldn't set them as final without also creating a constructor
that also accepted all instance variables in the extending classes
and, at that point, it was just going to be easier to create them as
separate classes. So I'll need to change that old `Post` class to a
`PostRetrieve` class and then we should be set.