I was originally planning to use `PostSent` for both creating and publishing posts but, just barely, the stuff sent for updating a post isn't a perfect subset of what's sent to create one so I'll just use a separate object.
It avoid the inevitable collision that'd result from having three
arguments to the constructor that're `String`s (fixed in the
`PostUpdate` class) and, if anything changes to either/or
individually, I'll be able to easily rectify. And it'll be less
confusing to the user. And not expand further the size of the new
`PostCreated` class.