Add the `Channel` object and the calls for it.
I opted just to make another class instead of just adding to the
`CrosspostInfo` class because that would mean it'd get serialized and
deserialized in very different ways (which would probably be alright
to pull off; I'm pretty sure I wouldn't ever have to send it to the
API other than for the crossposting parameter).

But mostly because the point of creating the `Enum` to designate each
channel ID was that you can only do Tumblr, Medium, or
Twitter (according to the documentation). To extend `CrosspostInfo`,
I'd (probably) have to either ditch the `Enum` since I don't see all
possible channels listed anywhere in the documentation yet (I haven't
made it fully all the way through, yet, though I've clocked ~75% or
more) or, provided I had a list of everything, I could make the `Enum`
exhaustive and just throw an `Exception` when it gets serialized if
one of the non-appropriate ones were sent for the crosspost parameter.

Since there's still a value in having a designated `Enum` for
crossposting and the above hasn't been found to be true (yet), we'll
just create a separate object.