I accidentally transcribed the deserializer for `ZonedDateTime` incorrectly, 'wasn't including the "Token " part of the "Authorization" `String` (though that doesn't seem to actually matter), and, in response to my moving `authToken` into `WriteFreelyAPIabstract` since it's used by both subclasses – anyway –, I figured I'd reference it directly instead of passing it through the method since the changes to `authToken` weren't getting tracked that way.
I think this is all I'm going to do for the night but I need to check
out possibly adding a deserializer for `ResponseWrapper` since error
codes aren't properly getting changed into proper objects (though near
every other iteration of the class gets recognized…; it might be the
underscore and I just need to use `@SerializedName`, like the
`Enum`s).

Also, response codes of 400 don't seem to be triggering the
authenticator, for some reason, which is killing `retrieveUser()`'s
usage (`publishPost()`, for example, returns a response code of 401,
when it fails). I'll need to keep in mind that, for something like
write.as, trying to call `publishPost()` while intending to be signed
in will fail since you can publish posts anonymously; I should
probably get the `login()` call working so people can do it
manually. I'll also need to check with Matt that trying to publish a
post with an outdated token will fail rather than gracefully fail and
publish anonymously. I suppose users of this could just check the
results and claim the post if that happens but the least amount of
manual work is the best.

That said…it works. It still really irritates me that every call in
`WriteFreelyAPI` is also redone, character for character, in
WriteFreelyAPIwithUser` but, eh. I still was able to reduce everything
that's shared down to the abstract class, at least, and both
subclasses are able to make use of them.